Make @vue/composition-api optional
NPM 7 attempts to automatically resolve peer dependencies and errors if it can't. This makes it difficult to use this package with Vue 3, since @vue/composition-api depends on Vue 2. This change brings vuefire in line with the README from vue-demi.

Pending this change, a workaround is to use --legacy-peer-deps when installing.