feat(firestore): allow custom converter
Close #608

BREAKING CHANGE: `options.serialize()` is replaced with `converter`. It
effectively has the same effect as calling `doc().withConverter()` or
`collection().withConverter()` but it allows to have a global converter
that is automatically applied to all snapshots. This custom converter
adds a non-enumerable `id` property for documents like the previous
`serialize` options. **If you were not using this option**, you don't
need to change anything.