feat: better defaults for wait and reset
BREAKING CHANGE: `wait` option now defaults to `true` to better align
with SSR. Similarly, the `reset` option now defaults to `false` to
better align with declarative usage of `useDocument()` and others. If
you want to keep the old behavior, you can still override the defaults
globally (refer to global options in the docs).