feat(database): rename `.value` to `$value` for primitive values
BREAKING CHANGE: when binding to a primitive value in RTDB, VueFire used
to create an object with a property `.value` for the primitive vaule
itself. The `.` in front forces to always use a bracket syntax
(`obj['.value']`) while the `$` doesn't, making its usage cleaner. The
`$value` and `id` property created in the case of primitives are also
**enumerable** properties. This should make things easier to debug.